Problem: #69 App architecture inefficiencies and overhead
Solution: moderator folder and page specific templates in pages folder. #69
