+++ imports/startup/both/routes.js
@@ -14,24 +14,24 @@ SubsCache.ready = function() {return true}
 if (Meteor.isClient) { // only import them if this code is being executed on client side
   import '../../ui/layouts/MainBody.html'
   import '../../ui/pages/changedCurrencies/changedCurrencies.js'
-  import '../../ui/pages/hashpower/addHashpower'
+  import '../../ui/pages/moderator/hashpower/addHashpower.js'
   import '../../ui/pages/allHashpower'
-  import '../../ui/pages/hashpower/allHashaverage'
-  import '../../ui/pages/flaggedUsers/flaggedUsers'
-  import '../../ui/pages/hashpower/flaggedHashpower'
+  import '../../ui/pages/moderator/hashpower/allHashaverage'
+  import '../../ui/pages/moderator/flaggedUsers/flaggedUsers'
+  import '../../ui/pages/moderator/hashpower/flaggedHashpower'
   import '../../ui/pages/compareCurrencies'
   import '../../ui/pages/userProfile'
   import '../../ui/pages/transactions'
-  import '../../ui/pages/returnedCurrencies.js'
-  import '../../ui/pages/moderatorDash.js'
-  import '../../ui/pages/ratings.js'
+  import '../../ui/pages/returnedCurrencies/returnedCurrencies.js'
+  import '../../ui/pages/moderator/moderatorDash/moderatorDash'
+  import '../../ui/pages/ratings/ratings.js'
   import '../../ui/pages/theme.html'
   import '../../ui/pages/communities'
   import '../../ui/pages/codebase'
   import '../../ui/pages/developers'
   import '../../ui/pages/editProfile'
-  import '../../ui/pages/questions'
-  import '../../ui/pages/bounties'
+  import '../../ui/pages/moderator/questions/questions'
+  import '../../ui/pages/bounties/bounties'
   import '../../ui/pages/addCoin'
 
   // New Layout
