+++ imports/ui/pages/bounties/HashrateAPI.html
