+++ imports/ui/pages/bounties/HashrateAPI.js
