--- imports/ui/pages/hashpower/allHashaverage.js
@@ -1,31 +1,0 @@
-import { Template } from 'meteor/templating';
-import { HashAverage, HashAlgorithm } from '/imports/api/indexDB.js'
-
-import './allHashaverage.html'
-
-Template.allHashaverage.onCreated(function() {
-	this.autorun(() => {
-		SubsCache.subscribe('hashaverage')
-		SubsCache.subscribe('hashalgorithm')
-	})
-})
-
-Template.allHashaverage.helpers({
-	hashaverage: () => HashAverage.find({}).fetch(),
-	hashAlgorithm: function() {
-		return (HashAlgorithm.findOne({
-			_id: this.algorithm
-		}) || {}).name || ''
-	},
-	average: function() {
-		return isNaN(this.average) ? 'N\\A' : `${this.average} W`
-	}
-})
-
-Template.allHashaverage.events({
-	'click #js-update': (event, templateInstance) => {
-		event.preventDefault()
-
-		Meteor.call('updateAverages', (err, data) => {})
-	}
-})
