+++ imports/ui/pages/moderator/hashpower/addHashpower.html
