+++ imports/ui/pages/moderator/hashpower/allHashaverage.html
