+++ imports/ui/pages/moderator/moderatorDash/approveWalletImage.js
