+++ imports/ui/pages/ratings/currencyChoice.html
