+++ imports/ui/pages/ratings/currencyChoices.html
