+++ imports/ui/pages/ratings/currencyChoices.js
