+++ imports/ui/pages/ratings/displayRatings.html
