+++ imports/ui/pages/ratings/ratings.html
